Added phpstan baseline at level 0
