+++ composer.json
@@ -85,6 +85,7 @@
         "moontoast/math": "^1.1",
         "orchestra/testbench-core": "^5.0",
         "pda/pheanstalk": "^4.0",
+        "phpstan/phpstan": "^0.12.18",
         "phpunit/phpunit": "^8.4|^9.0",
         "predis/predis": "^1.1.1",
         "symfony/cache": "^5.0"
