+++ tests/Container/Fixtures/ContainerConcreteStub.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Container\Fixtures;
+
+class ContainerConcreteStub
+{
+    //
+}
