+++ tests/Container/Fixtures/ContainerImplementationStub.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Container\Fixtures;
+
+class ContainerImplementationStub implements IContainerContractStub
+{
+    //
+}
