+++ tests/Container/Fixtures/ContainerInjectVariableStub.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Illuminate\Tests\Container\Fixtures;
+
+class ContainerInjectVariableStub
+{
+    public $something;
+
+    public function __construct(ContainerConcreteStub $concrete, $something)
+    {
+        $this->something = $something;
+    }
+}
