+++ tests/Container/Fixtures/IContainerContractStub.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Container\Fixtures;
+
+interface IContainerContractStub
+{
+    //
+}
