+++ tests/Database/Fixtures/EloquentTestPhoto.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures;
+
+use Illuminate\Database\Eloquent\Model as Eloquent;
+
+class EloquentTestPhoto extends Eloquent
+{
+    protected $table = 'photos';
+    protected $guarded = [];
+
+    public function imageable()
+    {
+        return $this->morphTo();
+    }
+}
