+++ tests/Events/Fixtures/ExampleEvent.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Events\Fixtures;
+
+class ExampleEvent
+{
+    //
+}
