+++ tests/Integration/Auth/ApiAuthenticationWithEloquentTest/ApiAuthenticationWithEloquentTest.php
