+++ tests/Integration/Database/EloquentBelongsToTest/EloquentBelongsToTest.php
