+++ tests/Support/Fixtures/CustomDateClass.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Illuminate\Tests\Support\Fixtures;
+
 class CustomDateClass
 {
     protected $original;
