+++ tests/Support/Fixtures/EventStub.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Support\Fixtures;
+
+class EventStub
+{
+    //
+}
